Start debugging something broken in IR impl
It appears the GT and possibly other instrs are not working as designed.
Added simple method to simpler debugger to print out what is executing
and this will need to be expanded for the other debuggers and also
probably more state should be printed like the stack.

This is not working but the problems are becoming more visible.