Jump removal
I noticed duplication of nodes in the ASG so I changed nodes to check to
make sure they do not already exist.  This will mildly mess up step #
since step 11 and step 90 could be the same node.  Debugging will just
show step 11 in this example whereas before it would line up with
what we see in chowder interp.  A reasonable tradeoff for less nodes.

In IR I converted > and != to be math ops and removed them as jumps
and branches.  In practice, the opaqueness of determining dp and cc
as a fixed value meants nearly none of these branches got removed.
It is trivial to optimize these new instrs to a copy.

I noticed swch can have both paths end up going to the same location
so I remove the branching logic in that case.