Add new native interpreter
Translate bbs instrs into a single mega instr which is replacing each
instr as a set of Ruby code.  The code is then eval'd to be the execute
method for the megainstr.  This interpreter removes all the overhead of
ipcs to calling a megainstr and performing bb transition.

Consolidated all the IR interpreters to DRY up the same copied code.